requirements.txt
azure-functions==1.0.3
flake8==3.3.0
apscheduler>=3.3.0
requests>=2.0.0,<3.0.0
ff_espn_api>=1.1.9